#2193 extract common method and register
